Use Wrapper for Modal for consistent experience
