+++ aquawatch_mobile_app/components/SettingsDropdown.tsx
@@ -1,17 +1,10 @@
 import { Picker } from '@react-native-picker/picker';
-import React, { useCallback, useEffect, useState } from 'react';
-import {
-    View,
-    Text,
-    Modal,
-    TouchableOpacity,
-    Platform,
-    Dimensions,
-    TouchableWithoutFeedback,
-} from 'react-native';
-import Animated, { useAnimatedStyle, useSharedValue, withTiming } from 'react-native-reanimated';
+import React, { useRef } from 'react';
+import { View, Text, TouchableOpacity, Platform } from 'react-native';
 
 import { useColorScheme } from '@/contexts/ColorSchemeContext';
+
+import { ModalWrapper, ModalWrapperRef } from './modals/ModalWrapper';
 
 export default function SettingsDropdown({
     label,
@@ -25,37 +18,7 @@ export default function SettingsDropdown({
     onSelect: (value: string) => void;
 }) {
     const { isDark } = useColorScheme();
-    const [modalVisible, setModalVisible] = useState(false);
-
-    const { height } = Dimensions.get('window');
-
-    const backdropOpacity = useSharedValue(0);
-    const modalTranslateY = useSharedValue(height); // Start off-screen
-
-    const backdropAnimatedStyle = useAnimatedStyle(() => ({
-        opacity: backdropOpacity.value,
-    }));
-
-    const modalAnimatedStyle = useAnimatedStyle(() => ({
-        transform: [{ translateY: modalTranslateY.value }],
-    }));
-
-    const openModal = useCallback(() => {
-        backdropOpacity.value = withTiming(1, { duration: 300 });
-        modalTranslateY.value = withTiming(0, { duration: 200 });
-    }, [backdropOpacity, modalTranslateY]);
-
-    const closeModal = useCallback(() => {
-        backdropOpacity.value = withTiming(0, { duration: 300 });
-        modalTranslateY.value = withTiming(height, { duration: 300 });
-        setModalVisible(false);
-    }, [backdropOpacity, height, modalTranslateY]);
-
-    useEffect(() => {
-        if (modalVisible) {
-            openModal();
-        }
-    }, [modalVisible, openModal]);
+    const modalRef = useRef<ModalWrapperRef>(null);
 
     // iOS: Show a button, open Picker in a modal
     if (Platform.OS === 'ios') {
@@ -79,7 +42,7 @@ export default function SettingsDropdown({
                     {label}
                 </Text>
                 <TouchableOpacity
-                    onPress={() => setModalVisible(true)}
+                    onPress={() => modalRef.current?.openModal()}
                     style={{
                         backgroundColor: isDark ? '#333333' : 'white',
                         paddingVertical: 8,
@@ -99,42 +62,15 @@ export default function SettingsDropdown({
                         {options.find((o) => o.value === value)?.label || label}
                     </Text>
                 </TouchableOpacity>
-                <Modal
-                    visible={modalVisible}
-                    transparent
-                    animationType="none"
-                    onRequestClose={() => setModalVisible(false)}>
-                    <TouchableWithoutFeedback onPress={closeModal}>
-                        <Animated.View
-                            style={[
-                                {
-                                    flex: 1,
-                                    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-                                },
-                                backdropAnimatedStyle,
-                            ]}
-                        />
-                    </TouchableWithoutFeedback>
-
-                    <Animated.View
-                        style={[
-                            {
-                                position: 'absolute',
-                                bottom: 0,
-                                height: '40%',
-                                width: '100%',
-                                backgroundColor: isDark ? '#1a202c' : '#f1f1f1',
-                                borderTopLeftRadius: 20,
-                                borderTopRightRadius: 20,
-                                padding: 20,
-                            },
-                            modalAnimatedStyle,
-                        ]}>
+                <ModalWrapper
+                    ref={modalRef}
+                    modalHeight={'40%'}
+                    body={
                         <Picker
                             selectedValue={value}
                             onValueChange={(itemValue) => {
                                 onSelect(String(itemValue));
-                                setModalVisible(false);
+                                modalRef.current?.closeModal();
                             }}
                             style={{
                                 color: isDark ? 'white' : 'black',
@@ -148,8 +84,8 @@ export default function SettingsDropdown({
                                 />
                             ))}
                         </Picker>
-                    </Animated.View>
-                </Modal>
+                    }
+                />
             </View>
         );
     }
